Add files via upload
A starbucks is one of the most well-known companies in the world. It strives to give his customers always the best service and the best experience. They have a mobile application in which the users can make orders online.The project aims at optimising the customers experience using the app through user's behaviour analysis.